Refactor patching to specific submodule (#2639)
* Create patching submodule

* Minor fix in docstring section header
